support 64-bit Int in WDL (using Long internally in WomInteger) #2685

+++ backend/src/main/scala/cromwell/backend/validation/ContinueOnReturnCodeValidation.scala
@@ -39,7 +39,7 @@ class ContinueOnReturnCodeValidation extends RuntimeAttributesValidation[Continu
     case WomBoolean(value) => ContinueOnReturnCodeFlag(value).validNel
     case WomString(value) if Try(value.toBoolean).isSuccess => ContinueOnReturnCodeFlag(value.toBoolean).validNel
     case WomString(value) if Try(value.toInt).isSuccess => ContinueOnReturnCodeSet(Set(value.toInt)).validNel
-    case WomInteger(value) => ContinueOnReturnCodeSet(Set(value)).validNel
+    case WomInteger(value) => ContinueOnReturnCodeSet(Set(value.toInt)).validNel
     case value@WomArray(_, seq) =>
       val errorOrInts: ErrorOr[List[Int]] = (seq.toList map validateInt).sequence[ErrorOr, Int]
       errorOrInts match {
